Refactor tensor and matrix operations for improved performance and consistency
- Updated Model class to use tensor_t instead of matrix_t for inputs and predictions.
- Modified Optimizer interface to accept tensor_t for gradient updates.
- Refactored SGD optimizer to handle tensor_t for weight updates and momentum calculations.
- Changed shape constructor to accept size_t instead of int for better type safety.
- Enhanced tensor class with new build method for initialization and added array-like access.
- Updated mnist dataset loading to utilize tensor_t for image and label storage.
- Adjusted CPU backend operations to support tensor operations and removed unnecessary comments.
- Implemented CUDA operations for tensor addition, subtraction, multiplication, and division.
- Removed deprecated matrix tests and adjusted tests to focus on tensor operations.
- Updated main application to utilize tensor_t for matrix multiplications and tests.
- Cleaned up code by removing unused test files and ensuring consistency across tensor operations.

+++ src/blust/include/blust/shape.hpp
@@ -15,7 +15,7 @@ public:
     typedef std::vector<size_t> dim_t;
 
     // Create a shape object with given inital dimensions as a initializer list
-    shape(std::initializer_list<int> dims) noexcept
+    shape(std::initializer_list<size_t> dims) noexcept
     {
         m_dims.reserve(dims.size());
         for (auto& d : dims)
